Add files via upload
First the code starts out by including some libraries, then defining two structures that I use, the neighbor and the City. The neighbor contains the name, distance from a parent, and the fare. The city class I used utilized accumulated distances and fares to associate with each city so that they could be stored effectively for the A* search. It will also allow storage of the proper heuristics to use within the search. The city class contains members for both a city with an empty name, (when queues are empty) and for a string when we have a name of a city. There is also an AddNeighbor function that essentially ties a neighbor to a city as well as its parameters. I prototype the two functions before the main program. Which runs BuildMap, then the search which does the task of moving along the path and printing out information. The BuildMap function uses the City Class and the AddNeighbor function to hardcode the map from the assignment and link all the cities together correctly. Once it creates a city, it just pushes it to the back of the city structure until the list is complete. Another tricky thing was to convert a neighbor city to an actual city node, instead of a neighbor node. To do this, there is a function called GetChildCity that just looks at the names of the neighbors and of all the cities and converts a neighbor into a city, so that it has the correct information. 

Finally the search function at line 245. First there is a lot of declaration of variables as it cycles through various city structures when looking through nodes. Some of the variables are unused becuase I was trying to use them differently. The Search Function contains all the necessary code for the Greedy Search and the A* search. The vector and deque methods are essential in utilizing the queueing system. Lines 296 through 309 basically search for the Starting system based off a string provided. Obviously we have Los Angeles so it will find the correct city information for Los Angeles and put that in the frontier. Then it runs the respective code for the search.