Add files via upload
Well, this wasn't the most fun thing to debug, especially because I did it in nested loops with multiple loops within loops. But anyways the code starts out by defining to classes, the board, and the piece. The Board is capable of storing: the layout which is a 8x8 char array, the heuristic value associated with it, as well as the pieces that were moved on it to get it to it's specific state. The Piece class stores the information about the piece such as location,  kingsmanship and the symbol associated with it. These will be useful for the moves functions. 

After that, I prototype all the functions, many of which go with finding specific moves for a given piece. Finally the main function takes an input and stores it in an 8x8 char array to be used as the input game board. After the 8x8 char's have been entered, you enter a depth number from 1 to 4. This is how deep the minimax search goes. At line 79 is the actual minmax search function which starts out by initilizing and definine variable and deques and iterators. Essentially, the search first looks at the input game board and stores all of A's pieces in a deque using the getApieces function. Once it has all the pieces, it cycles through each piece and generates a move associated with that piece. For each piece move it generates a new gameboard and stores that in another queue. One interesting note is that if there are jumps available, the code ignores regular moves for any of the pieces, becuase in checkers a player must jump if it is available. Once it has all the possible moves for a given game board, it starts to branch the moves out with a while loop. That is the main concept of the code. It is repeated for B's pieces and moves, then A's, then B's... When the search gets to the given depth such as at line 180, it evaluates the particular game board it is looking at and gives it the correct heuristic. A search depth of 1 would only look at how beneficial one move from A would be. If we increase the search depth we can start to expand the moves of B. By applying the alpha beta pruning algorithm we can evaluate particular branches of the tree and see how they favor against branches that have already been expanded and evaluated. We must pass the correct alpha and beta values through the tree in order to prune the correct branch. This requires many conditional statements that correspond with the maximum heuristic values and minimum. Min wants the heuristic to be small and Max wants it to be large. It has to check the sets of heuristics for the largest of the smallest of the largest of the smallest values ( for a depth of 4) That is what the code does in order to give it an optimal starting move for A. The pruning section is used in order to skip certain game states that do not yield any change in the alpha beta values as seen in lines 251 through 261. This is done at each ply of the game. I also have a check to see if there are possible moves or not by seeing if there are gamestates available for a given layout as seen in lines 317 through 323. During each loop it checks alpha and beta values for any possible updates.

Lines 493+ are all the functions that I used. The pieces functions just iterates through the array and stores info about the piece it's at.

The getJumps functions looks in the location around a certain piece to see if there are available jumps and changes the board array by reference and stores it in a queue. The getMoves functions work in a similar way as they have to look at the layout around them in order to be able to move. Because I wrote the functions as forward and backward, I didn't bother retyping out new functions for B and instead just used similar jump functions (because they compare piece symbols) Finally the heuristic function evaluates the board.